Use caption of the lightgallery selector.
$('#images').lightGallery({
    selector: '.image'
}); 
<div id="images">
  <div class="image" data-src="bigImg.jpg" data-sub-html=".caption" data-sub-html-self="true">
     <img src="img.jpg"/>
     <div class="caption">
          <h3>Image Title</h3>
          <p>Image Text</p>
     </div>
  </div>
  <div class="image" data-src="bigImg2.jpg" data-sub-html=".caption" data-sub-html-self="true">
     <img src="img2.jpg"/>
     <div class="caption">
          <h3>Image Title 2</h3>
          <p>Image Text 2</p>
     </div>
  </div>
  ....
</div

With this change it is possible to use the "sub-html" of the current element therefore you don't have to search a specific selector within the whole dom and without a unique identifier. 
Benefits: no unique identifier needed, caption is within the lightgallery selector, faster because you don't have to find the element within the whole DOM

Thx for the nice library!